Autenticação de usuários nos apps

+++ apps/products/views.py
@@ -5,48 +5,62 @@ from .models import Item
 
 
 def item_dashboard(request):
-    itens = Item.objects.order_by("name")
-    paginator = Paginator(itens, 30)
-    page = request.GET.get("page")
-    items_per_page = paginator.get_page(page)
-    package = {"items": items_per_page}
-    return render(request, "items/item_dashboard.html", package)
-
+    if request.user.is_authenticated:
+        itens = Item.objects.order_by("name")
+        paginator = Paginator(itens, 30)
+        page = request.GET.get("page")
+        items_per_page = paginator.get_page(page)
+        package = {"items": items_per_page}
+        return render(request, "items/item_dashboard.html", package)
+    else:
+        return redirect("index")
 
 def get_item(request, item_id):
-    item = get_object_or_404(Item, pk=item_id)
-    item_to_show = {"item": item}
-    return render(request, "items/item.html", item_to_show)
-
+    if request.user.is_authenticated:
+        item = get_object_or_404(Item, pk=item_id)
+        item_to_show = {"item": item}
+        return render(request, "items/item.html", item_to_show)
+    else:
+        return redirect("index")
 
 def create_item(request):
-    if request.method == "POST":
-        name = request.POST["item_name"]
-        price = request.POST["item_price"]
-        item = Item.objects.create(name=name, price=price)
-        item.save()
-        return redirect("item_dashboard")
+    if request.user.is_authenticated:
+        if request.method == "POST":
+            name = request.POST["item_name"]
+            price = request.POST["item_price"]
+            item = Item.objects.create(name=name, price=price)
+            item.save()
+            return redirect("item_dashboard")
+        else:
+            return render(request, "items/create_item.html")
     else:
-        return render(request, "items/create_item.html")
-
+        return redirect("index")
 
 def delete_item(request, item_id):
-    item = get_object_or_404(Item, pk=item_id)
-    item.delete()
-    return redirect("item_dashboard")
-
+    if request.user.is_authenticated:
+        item = get_object_or_404(Item, pk=item_id)
+        item.delete()
+        return redirect("item_dashboard")
+    else:
+        return redirect("index")
 
 def edit_item(request, item_id):
-    item = get_object_or_404(Item, pk=item_id)
-    item_to_edit = {"item": item}
-    return render(request, "items/edit_item.html", item_to_edit)
-
+    if request.user.is_authenticated:
+        item = get_object_or_404(Item, pk=item_id)
+        item_to_edit = {"item": item}
+        return render(request, "items/edit_item.html", item_to_edit)
+    else:
+        return redirect("index")
 
 def update_item(request):
-    if request.method == "POST":
-        item_id = request.POST["item_id"]
-        item = Item.objects.get(pk=item_id)
-        item.name = request.POST["item_name"]
-        item.price = request.POST["item_price"]
-        item.save()
-    return redirect("item_dashboard")
+    if request.user.is_authenticated:
+        if request.method == "POST":
+            item_id = request.POST["item_id"]
+            item = Item.objects.get(pk=item_id)
+            item.name = request.POST["item_name"]
+            item.price = request.POST["item_price"]
+            item.save()
+        return redirect("item_dashboard")
+    else:
+        return redirect("index")
+
